小程序二维码接口代码优化
