+++ weixin-java-miniapp/src/main/java/cn/binarywang/wx/miniapp/bean/WxMaWxcode.java
@@ -1,6 +1,5 @@
 package cn.binarywang.wx.miniapp.bean;
 
-import cn.binarywang.wx.miniapp.api.WxMaQrcodeService;
 import cn.binarywang.wx.miniapp.util.json.WxMaGsonBuilder;
 import com.google.gson.annotations.SerializedName;
 
@@ -19,7 +18,7 @@ public class WxMaWxcode extends WxMaQrcodeWrapper implements Serializable {
   private boolean autoColor = true;
 
   @SerializedName("line_color")
-  private WxMaQrcodeService.LineColor lineColor = new WxMaQrcodeService.LineColor("0", "0", "0");
+  private WxMaCodeLineColor lineColor = new WxMaCodeLineColor("0", "0", "0");
 
   public static WxMaWxcode fromJson(String json) {
     return WxMaGsonBuilder.create().fromJson(json, WxMaWxcode.class);
@@ -53,11 +52,11 @@ public class WxMaWxcode extends WxMaQrcodeWrapper implements Serializable {
     this.autoColor = autoColor;
   }
 
-  public WxMaQrcodeService.LineColor getLineColor() {
+  public WxMaCodeLineColor getLineColor() {
     return lineColor;
   }
 
-  public void setLineColor(WxMaQrcodeService.LineColor lineColor) {
+  public void setLineColor(WxMaCodeLineColor lineColor) {
     this.lineColor = lineColor;
   }
