+++ weixin-java-miniapp/src/main/java/cn/binarywang/wx/miniapp/bean/WxMaWxcodeLimit.java
@@ -1,6 +1,5 @@
 package cn.binarywang.wx.miniapp.bean;
 
-import cn.binarywang.wx.miniapp.api.WxMaQrcodeService;
 import cn.binarywang.wx.miniapp.util.json.WxMaGsonBuilder;
 import com.google.gson.annotations.SerializedName;
 
@@ -20,7 +19,7 @@ public class WxMaWxcodeLimit extends WxMaQrcodeWrapper implements Serializable {
   private boolean autoColor = true;
 
   @SerializedName("line_color")
-  private WxMaQrcodeService.LineColor lineColor = new WxMaQrcodeService.LineColor("0", "0", "0");
+  private WxMaCodeLineColor lineColor = new WxMaCodeLineColor("0", "0", "0");
 
   public static WxMaWxcodeLimit fromJson(String json) {
     return WxMaGsonBuilder.create().fromJson(json, WxMaWxcodeLimit.class);
@@ -58,11 +57,11 @@ public class WxMaWxcodeLimit extends WxMaQrcodeWrapper implements Serializable {
     this.autoColor = autoColor;
   }
 
-  public WxMaQrcodeService.LineColor getLineColor() {
+  public WxMaCodeLineColor getLineColor() {
     return lineColor;
   }
 
-  public void setLineColor(WxMaQrcodeService.LineColor lineColor) {
+  public void setLineColor(WxMaCodeLineColor lineColor) {
     this.lineColor = lineColor;
   }
 }
